requirements.txt
PyYAML
schedule
docker